django/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
graphene-django>=3.2,<4.0
django-filter>=24.0,<25.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
dj-database-url>=2.1,<3.0
